add casts to suppress warnings.

+++ size.c
@@ -28,15 +28,15 @@
 
 int main(int argc, char *argv[])
 {
-  printf("sizeof(struct sockaddr_un) = %d\n", sizeof(struct sockaddr_un));
-  printf("sizeof(struct sockaddr_un.sun_family) = %d\n", FIELD_SIZE(struct sockaddr_un, sun_family));
-  printf("sizeof(struct sockaddr_un.sun_path) = %d\n", FIELD_SIZE(struct sockaddr_un, sun_path));
-  printf("offsetof(struct sockaddr_un.sun_family) = %d\n", offsetof(struct sockaddr_un, sun_family));
-  printf("offsetof(struct sockaddr_un.sun_path) = %d\n", offsetof(struct sockaddr_un, sun_path));
+  printf("sizeof(struct sockaddr_un) = %d\n", (int)sizeof(struct sockaddr_un));
+  printf("sizeof(struct sockaddr_un.sun_family) = %d\n", (int)FIELD_SIZE(struct sockaddr_un, sun_family));
+  printf("sizeof(struct sockaddr_un.sun_path) = %d\n", (int)FIELD_SIZE(struct sockaddr_un, sun_path));
+  printf("offsetof(struct sockaddr_un.sun_family) = %d\n", (int)offsetof(struct sockaddr_un, sun_family));
+  printf("offsetof(struct sockaddr_un.sun_path) = %d\n", (int)offsetof(struct sockaddr_un, sun_path));
 
-  printf("sizeof(struct sockaddr_storage) = %d\n", sizeof(struct sockaddr_storage));
-  printf("sizeof(struct sockaddr_storage.ss_family) = %d\n", FIELD_SIZE(struct sockaddr_storage, ss_family));
-  printf("offsetof(struct sockaddr_storage.sun_family) = %d\n", offsetof(struct sockaddr_storage, ss_family));
+  printf("sizeof(struct sockaddr_storage) = %d\n", (int)sizeof(struct sockaddr_storage));
+  printf("sizeof(struct sockaddr_storage.ss_family) = %d\n", (int)FIELD_SIZE(struct sockaddr_storage, ss_family));
+  printf("offsetof(struct sockaddr_storage.sun_family) = %d\n", (int)offsetof(struct sockaddr_storage, ss_family));
 
   return EXIT_SUCCESS;
 }
